Separate task editor css into different files
