+++ frontend/src/components/FloatingTaskEditor/TaskEditor.css
@@ -1,0 +1,147 @@
+/*
+ * --------------------------------------------------------------------------------
+ * Part 1: Parent Component
+ * --------------------------------------------------------------------------------
+ */
+
+.TaskEditor {
+  z-index: 7;
+  top: 0;
+  left: 0;
+  width: 300px;
+  height: auto;
+  color: white;
+  font-size: 12px;
+  box-shadow: 0 5px 12px 0 rgba(0, 0, 0, 0.3);
+}
+
+.TaskEditorSubmitButtonRow {
+  display: flex;
+  padding: 0.5em;
+}
+
+.TaskEditorSaveButton {
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  flex: 0;
+  width: auto;
+  height: 2em;
+  border-radius: 2px;
+  border: solid 2px white;
+  background-color: rgba(255, 255, 255, 0.15);
+}
+
+.TaskEditorSaveButtonText {
+  font-size: 12px;
+  color: white;
+  margin: 0 0.5em;
+}
+
+/*
+ * --------------------------------------------------------------------------------
+ * Part 2: Common Components for Both Children Components
+ * --------------------------------------------------------------------------------
+ */
+
+.TaskEditorHeader {
+  font-size: 1em;
+}
+
+.TaskEditorFlexibleContainer {
+  position: relative;
+  display: flex;
+  margin: 1em 0.5em;
+  align-content: center;
+  align-items: center;
+}
+
+.TaskEditorFlexiblePadding {
+  flex: 1 1 auto;
+}
+
+.TaskEditorFlexibleInput {
+  flex: 1 1 auto;
+}
+
+.TaskEditorFlexibleInput > input, .TaskEditorFlexibleInput > div > input {
+  margin-left: 0.5em !important;
+  padding: 0.5em 0 !important;
+  border-radius: 0 !important;
+  background-color: transparent !important;
+  line-height: 1em !important;
+  border: none !important;
+  color: white !important;
+  border-bottom: 1px solid transparent !important;
+}
+
+.TaskEditorFlexibleInput > input:focus-within {
+  border-bottom-color: white !important;
+}
+
+.TaskEditorFlexibleInput > input::placeholder,
+.TaskEditorFlexibleInput > input:focus-within::placeholder {
+  color: white !important;
+}
+
+.TaskEditorCheckBox {
+  margin-left: 8px;
+  margin-right: 0;
+}
+
+/*
+ * --------------------------------------------------------------------------------
+ * Part 3: Components for Main Task Editor
+ * --------------------------------------------------------------------------------
+ */
+
+.TaskEditorIconButton {
+  margin-left: 1em !important;
+}
+
+.TaskEditorTag {
+  margin-left: 0.5em;
+  color: white;
+}
+
+.TaskEditorTagLabel {
+  font-size: 1em;
+}
+
+.TaskEditorTag input[type="checkbox"] {
+  position: absolute;
+  left: -2px;
+  width: 22px;
+  height: 22px;
+  opacity: 0;
+  margin: 0;
+  cursor: pointer;
+}
+
+.TaskEditorTagEditor {
+  position: absolute;
+  right: 0;
+  left: calc(100% - 150px);
+  top: 0;
+  z-index: 4;
+  color: #333333;
+  background: white;
+}
+
+.TaskEditorCalendar {
+  position: absolute;
+  right: 0;
+  left: calc(100% - 282px);
+  top: 36px;
+  z-index: 4;
+}
+
+/*
+ * --------------------------------------------------------------------------------
+ * Part 4: Components for Sub-Task Editor
+ * --------------------------------------------------------------------------------
+ */
+
+.TaskEditorSubTasksIndentedContainer {
+  margin-left: 2em;
+}
